Added individual routes for each friend
